+++ examples/generated-modules/src/main/java/remote/ClientProvider.kt
@@ -17,14 +17,15 @@
 
 package it.czerwinski.android.hilt.examples.generated.remote
 
+import android.util.Log
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.android.Android
-import io.ktor.client.features.json.JsonFeature
-import io.ktor.client.features.json.serializer.KotlinxSerializer
-import io.ktor.client.features.logging.ANDROID
-import io.ktor.client.features.logging.LogLevel
-import io.ktor.client.features.logging.Logger
-import io.ktor.client.features.logging.Logging
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.client.plugins.logging.MessageLengthLimitingLogger
+import io.ktor.serialization.kotlinx.json.json
 import it.czerwinski.android.hilt.annotations.FactoryMethod
 import javax.inject.Singleton
 
@@ -39,12 +40,23 @@ object ClientProvider {
             connectTimeout = TIMEOUT
             socketTimeout = TIMEOUT
         }
-        install(JsonFeature) {
-            serializer = KotlinxSerializer()
+        install(ContentNegotiation) {
+            json()
         }
         install(Logging) {
-            logger = Logger.ANDROID
+            logger = MessageLengthLimitingLogger(delegate = AndroidLogger)
             level = LogLevel.BODY
+        }
+    }
+
+    private object AndroidLogger : Logger {
+
+        private const val TAG = "HttpClient"
+
+        override fun log(message: String) {
+            for (line in message.lines()) {
+                Log.d(TAG, line)
+            }
         }
     }
 }